asdfasdf

+++ src/complex.c
@@ -1,6 +1,6 @@
 #include "complex.h"
 
-Complex *complex_new(float r, float i) {
+Complex *complex_new(double r, double i) {
 	Complex *self = (Complex *)malloc(sizeof(Complex));
 
 	self->real = r;
@@ -9,16 +9,16 @@ Complex *complex_new(float r, float i) {
 	return self;
 }
 
-float complex_mag(Complex *self) {
+double complex_mag(Complex *self) {
 	assert(self);
 
-	return sqrtf(complex_magSqr(self));
+	return sqrt(complex_magSqr(self));
 }
 
-float complex_magSqr(Complex *self) {
+double complex_magSqr(Complex *self) {
 	assert(self);
 
-	return powf(self->real, 2) + powf(self->imaginary, 2);
+	return pow(self->real, 2) + pow(self->imaginary, 2);
 }
 
 int complex_mandelbrot_iterations(Complex *self, int maxIter) {
@@ -29,7 +29,7 @@ int complex_mandelbrot_iterations(Complex *self, int maxIter) {
 	z.imaginary = 0;
 
 	for (int i = 0; i < maxIter; ++i) {
-		float tempReal = z.real;
+		double tempReal = z.real;
 		z.real = (z.real * z.real) - (z.imaginary * z.imaginary);
 		z.imaginary *= 2 * tempReal;
 		z.real += self->real;
@@ -43,13 +43,13 @@ int complex_mandelbrot_iterations(Complex *self, int maxIter) {
 	return maxIter - 1;
 }
 
-int mandelbrot_iterations(float real, float imaginary, int maxIter) {
+int mandelbrot_iterations(double real, double imaginary, int maxIter) {
 	Complex z;
 	z.real = 0;
 	z.imaginary = 0;
 
 	for (int i = 0; i < maxIter; ++i) {
-		float tempReal = z.real;
+		double tempReal = z.real;
 		z.real = (z.real * z.real) - (z.imaginary * z.imaginary);
 		z.imaginary *= 2 * tempReal;
 		z.real += real;
