+++ src/complex.h
@@ -6,14 +6,14 @@
 #include <stdlib.h>
 
 typedef struct complex_t {
-	float real;
-	float imaginary;
+	double real;
+	double imaginary;
 } Complex;
 
-Complex *complex_new(float r, float i);
-float complex_mag(Complex *self);
-float complex_magSqr(Complex *self);
+Complex *complex_new(double r, double i);
+double complex_mag(Complex *self);
+double complex_magSqr(Complex *self);
 int complex_mandelbrot_iterations(Complex *self, int maxIter);
-int mandelbrot_iterations(float real, float imaginary, int maxIter);
+int mandelbrot_iterations(double real, double imaginary, int maxIter);
 
 #endif
